Restructure theorem rankers, add comments
