+++ src/core/contextTheoremRanker/actualRankers/distanceContextTheoremsRanker.ts
@@ -1,7 +1,7 @@
-import { Theorem } from "../../coqParser/parsedTypes";
-import { CompletionContext } from "../completionGenerationContext";
+import { Theorem } from "../../../coqParser/parsedTypes";
+import { CompletionContext } from "../../completionGenerationContext";
 
-import { ContextTheoremsRanker } from "./contextTheoremsRanker";
+import { ContextTheoremsRanker } from "../contextTheoremsRanker";
 
 export class DistanceContextTheoremsRanker implements ContextTheoremsRanker {
     readonly needsUnwrappedNotations = false;
