+++ src/core/contextTheoremRanker/actualRankers/euclidContextTheoremRanker.ts
@@ -1,19 +1,16 @@
-import { Theorem } from "../../coqParser/parsedTypes";
-import { CompletionContext } from "../completionGenerationContext";
+import { Theorem } from "../../../coqParser/parsedTypes";
+import { CompletionContext } from "../../completionGenerationContext";
 
-import { ContextTheoremsRanker } from "./contextTheoremsRanker";
-import { goalAsTheoremString } from "./tokenUtils";
+import { ContextTheoremsRanker } from "../contextTheoremsRanker";
+import { goalAsTheoremString } from "../utils/tokenUtils";
 
 /**
  * Ranks theorems based on how similar their statements are to
  * the current goal context. Metric is calculated on the
  * concatenated hypothesis and conclusion.
  *
- * ```J(A, B) = |A ∩ B| / |A ∪ B|```
  */
-export class JaccardIndexContextTheoremsRanker
-    implements ContextTheoremsRanker
-{
+export class EuclidContextTheoremsRanker implements ContextTheoremsRanker {
     readonly needsUnwrappedNotations = true;
 
     rankContextTheorems(
@@ -23,7 +20,7 @@ export class JaccardIndexContextTheoremsRanker
         const goal = completionContext.proofGoal;
         const goalTheorem = goalAsTheoremString(goal);
 
-        const jaccardIndex = (theorem: Theorem): number => {
+        const euclid = (theorem: Theorem): number => {
             const completionTokens = goalTheorem
                 .split(" ")
                 .filter(
@@ -43,9 +40,9 @@ export class JaccardIndexContextTheoremsRanker
 
             const union = new Set([...completionTokens, ...theoremTokens]);
 
-            return intersection.length / union.size;
+            return Math.sqrt(intersection.length - union.size);
         };
 
-        return theorems.sort((a, b) => jaccardIndex(b) - jaccardIndex(a));
+        return theorems.sort((a, b) => euclid(b) - euclid(a));
     }
 }
