+++ src/core/contextTheoremRanker/actualRankers/jaccardIndexContextTheoremsRanker.ts
@@ -1,17 +1,19 @@
-import { Theorem } from "../../coqParser/parsedTypes";
-import { CompletionContext } from "../completionGenerationContext";
+import { Theorem } from "../../../coqParser/parsedTypes";
+import { CompletionContext } from "../../completionGenerationContext";
 
-import { ContextTheoremsRanker } from "./contextTheoremsRanker";
-import { goalAsTheoremString } from "./tokenUtils";
+import { ContextTheoremsRanker } from "../contextTheoremsRanker";
+import { goalAsTheoremString } from "../utils/tokenUtils";
 
 /**
  * Ranks theorems based on how similar their statements are to
  * the current goal context. Metric is calculated on the
  * concatenated hypothesis and conclusion.
  *
- * ```cosine(A, B) = |A ∩ B| / sqrt(|A| * |B|)```
+ * ```J(A, B) = |A ∩ B| / |A ∪ B|```
  */
-export class CosineContextTheoremsRanker implements ContextTheoremsRanker {
+export class JaccardIndexContextTheoremsRanker
+    implements ContextTheoremsRanker
+{
     readonly needsUnwrappedNotations = true;
 
     rankContextTheorems(
@@ -21,7 +23,7 @@ export class CosineContextTheoremsRanker implements ContextTheoremsRanker {
         const goal = completionContext.proofGoal;
         const goalTheorem = goalAsTheoremString(goal);
 
-        const cosine = (theorem: Theorem): number => {
+        const jaccardIndex = (theorem: Theorem): number => {
             const completionTokens = goalTheorem
                 .split(" ")
                 .filter(
@@ -39,12 +41,11 @@ export class CosineContextTheoremsRanker implements ContextTheoremsRanker {
                 theoremTokens.includes(token)
             );
 
-            return (
-                intersection.length /
-                Math.sqrt(completionTokens.length * theoremTokens.length)
-            );
+            const union = new Set([...completionTokens, ...theoremTokens]);
+
+            return intersection.length / union.size;
         };
 
-        return theorems.sort((a, b) => cosine(b) - cosine(a));
+        return theorems.sort((a, b) => jaccardIndex(b) - jaccardIndex(a));
     }
 }
