+++ src/core/contextTheoremRanker/actualRankers/randomContextTheoremsRanker.ts
@@ -1,7 +1,7 @@
-import { Theorem } from "../../coqParser/parsedTypes";
-import { CompletionContext } from "../completionGenerationContext";
+import { Theorem } from "../../../coqParser/parsedTypes";
+import { CompletionContext } from "../../completionGenerationContext";
 
-import { ContextTheoremsRanker } from "./contextTheoremsRanker";
+import { ContextTheoremsRanker } from "../contextTheoremsRanker";
 
 export class RandomContextTheoremsRanker implements ContextTheoremsRanker {
     readonly needsUnwrappedNotations = false;
