+++ src/core/contextTheoremRanker/actualRankers/weightedJaccardIndexTheoremRanker.ts
@@ -1,8 +1,8 @@
-import { Theorem } from "../../coqParser/parsedTypes";
-import { CompletionContext } from "../completionGenerationContext";
+import { Theorem } from "../../../coqParser/parsedTypes";
+import { CompletionContext } from "../../completionGenerationContext";
 
-import { ContextTheoremsRanker } from "./contextTheoremsRanker";
-import { goalAsTheoremString } from "./tokenUtils";
+import { ContextTheoremsRanker } from "../contextTheoremsRanker";
+import { goalAsTheoremString } from "../utils/tokenUtils";
 
 /**
  * Ranks theorems based on how similar their statements are to
