+++ src/core/contextTheoremRanker/actualRankers/сosineContextTheoremRanker.ts
@@ -1,16 +1,17 @@
-import { Theorem } from "../../coqParser/parsedTypes";
-import { CompletionContext } from "../completionGenerationContext";
+import { Theorem } from "../../../coqParser/parsedTypes";
+import { CompletionContext } from "../../completionGenerationContext";
 
-import { ContextTheoremsRanker } from "./contextTheoremsRanker";
-import { goalAsTheoremString } from "./tokenUtils";
+import { ContextTheoremsRanker } from "../contextTheoremsRanker";
+import { goalAsTheoremString } from "../utils/tokenUtils";
 
 /**
  * Ranks theorems based on how similar their statements are to
  * the current goal context. Metric is calculated on the
  * concatenated hypothesis and conclusion.
  *
+ * ```cosine(A, B) = |A ∩ B| / sqrt(|A| * |B|)```
  */
-export class EuclidContextTheoremsRanker implements ContextTheoremsRanker {
+export class CosineContextTheoremsRanker implements ContextTheoremsRanker {
     readonly needsUnwrappedNotations = true;
 
     rankContextTheorems(
@@ -20,7 +21,7 @@ export class EuclidContextTheoremsRanker implements ContextTheoremsRanker {
         const goal = completionContext.proofGoal;
         const goalTheorem = goalAsTheoremString(goal);
 
-        const euclid = (theorem: Theorem): number => {
+        const cosine = (theorem: Theorem): number => {
             const completionTokens = goalTheorem
                 .split(" ")
                 .filter(
@@ -38,11 +39,12 @@ export class EuclidContextTheoremsRanker implements ContextTheoremsRanker {
                 theoremTokens.includes(token)
             );
 
-            const union = new Set([...completionTokens, ...theoremTokens]);
-
-            return Math.sqrt(intersection.length - union.size);
+            return (
+                intersection.length /
+                Math.sqrt(completionTokens.length * theoremTokens.length)
+            );
         };
 
-        return theorems.sort((a, b) => euclid(b) - euclid(a));
+        return theorems.sort((a, b) => cosine(b) - cosine(a));
     }
 }
