+++ src/core/contextTheoremRanker/utils/tokenUtils.ts
@@ -1,6 +1,6 @@
-import { Goal, PpString } from "../../coqLsp/coqLspTypes";
+import { Goal, PpString } from "../../../coqLsp/coqLspTypes";
 
-import { hypToString } from "../exposedCompletionGeneratorUtils";
+import { hypToString } from "../../exposedCompletionGeneratorUtils";
 
 export function goalAsTheoremString(proofGoal: Goal<PpString>): string {
     const auxTheoremConcl = proofGoal?.ty;
